Ensure canvas is a part of the DOM tree
In case of external DOM manipulation - canvas will continue live in memory, so new canvas will not be created but topbar will not be visible. Example use-case is if `body` tag was replaced by htmx-boost. Same can be simulated by other scenarios which result canvas to be removed from the DOM.